Make zxcvbn a peer dependency
The zxcvbn package is accounting for a majority portion of the total asset
size in the Playbook project. As a regular dependency, this asset is
loaded on every page that includes Playbook in Nitro. Declaring it as
a peer dependency allows it to only be loaded on pages that need it,
which will as a result reduce the overall asset size that gets loaded
on each page.